requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
fastapi==0.99.1
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pydantic==1.10.10
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
